extracted the tests for `DropwizardPlatform`.
